+++ rust-src/widget/text_editor.rs
@@ -5,7 +5,8 @@ use iced::widget::{text_editor, TextEditor};
 use iced::{Background, Border, Color, Length, Padding};
 use text_editor::{Action, Content, Status, Style};
 
-use super::{read_c_string, ElementPtr, IcedMessage};
+use super::{ElementPtr, IcedMessage};
+use crate::ffi::read_c_string;
 
 type SelfPtr = *mut TextEditor<'static, PlainText, IcedMessage>;