+++ src/Iced/Widget/Canvas/FramePtr.hs
@@ -1,6 +1,0 @@
-module Iced.Widget.Canvas.FramePtr where
-
-import Foreign
-
-data NativeFrame
-type FramePtr = Ptr NativeFrame