+++ CHANGELOG.md
@@ -7,3 +7,5 @@
 * Removed dependency on TemplateHaskell
 * Renamed Command to Task
 * Added mouseArea widget
+* Renamed FrameAction to Action
+* Added image support for canvas